refactor: split memory/hub.rs into transport and domain
hub.rs said "Remote, shared memory tiers" and held six concerns: what a Memory is (local
included), repo identity, state classification, user-facing messages, credentials, and
repo lifecycle. Nothing was wrong locally — there was simply nowhere else for any of it
to go, which is how each command ended up inventing its own answer to "what state is
this memory in".

The line drawn here is transport vs domain:

- src/memory/hub.rs keeps transport only — the client, the token, repo identity
  (parse_hf, is_remote_shorthand) and repo lifecycle (whoami, create_dataset_repo). It
  now has no crate:: dependencies at all: a leaf that knows nothing about memories.
- src/memory.rs — until now just module declarations — becomes the domain: Memory and
  its parse/resolve/open/label, check_compat, the state signals (Reachability,
  remote_reachability, dataset_absent) and the three messages they render as. It calls
  down into hub for the Hub work.

So `hub::Memory` becomes `memory::Memory` throughout, and the tests split along the same
line: token and parse_hf tests stay with transport, Memory/check_compat/dataset_absent
tests move to the domain.

No logic changes. This is the file layout the state consolidation needs: one module that
owns "what state is a memory in", with everything that answers it already inside.
