+++ benchmark/bench_index.rs
@@ -17,7 +17,7 @@ use anyhow::Result;
 use arrow_array::{Array, StringArray};
 use clap::Parser;
 use funes::commands::index::run_index;
-use funes::memory::hub::Memory;
+use funes::memory::Memory;
 use futures::TryStreamExt;
 use std::collections::HashSet;
 use std::path::PathBuf;
