+++ src/commands/recall.rs
@@ -8,7 +8,7 @@ use super::curate;
 use crate::chunk;
 use crate::inference::{self, Embedder, Reranker};
 use crate::memory::dataset;
-use crate::memory::hub::{self, Memory, Reachability};
+use crate::memory::{self, Memory, Reachability};
 use crate::traces::harness::Harness;
 use anyhow::{anyhow, Context, Result};
 use arrow_array::{Float32Array, Int64Array, RecordBatch, StringArray, UInt64Array};
@@ -152,9 +152,9 @@ enum ReadOutcome {
 /// this; the remote-state classification and messages come from `hub`.
 async fn open_for_read(memory: &Memory) -> Result<ReadOutcome> {
     if let Memory::Remote { uri } = memory {
-        match hub::remote_reachability(uri).await {
+        match memory::remote_reachability(uri).await {
             Reachability::Offline => return Ok(ReadOutcome::Offline),
-            Reachability::Missing => return Err(hub::missing_remote(uri)),
+            Reachability::Missing => return Err(memory::missing_remote(uri)),
             Reachability::Ok => {}
         }
     }
@@ -163,16 +163,16 @@ async fn open_for_read(memory: &Memory) -> Result<ReadOutcome> {
         // The default local memory with no dataset yet → NoIndex (a clear "run funes add" error
         // below). Gated on `dataset_absent` so a real open failure (permissions, corruption, an
         // incompatible schema) isn't masked as "no index" — it falls through to surface as itself.
-        Err(e) if memory.is_default_local() && hub::dataset_absent(&e) => Ok(ReadOutcome::NoIndex),
+        Err(e) if memory.is_default_local() && memory::dataset_absent(&e) => Ok(ReadOutcome::NoIndex),
         // The Hub refused the read on auth (401/403): a clear message beats lance's opendal dump.
         Err(e) if is_auth_error(&e) => match memory {
-            Memory::Remote { uri } => Err(hub::unauthorized_remote(uri)),
+            Memory::Remote { uri } => Err(memory::unauthorized_remote(uri)),
             Memory::Local { .. } => Err(e),
         },
         // Opened to nothing: a reachable remote never pushed to, or a local path with no dataset.
         // Either way, a clear message beats lance's internal path error.
-        Err(e) if hub::dataset_absent(&e) => match memory {
-            Memory::Remote { uri } => Err(hub::empty_remote(uri)),
+        Err(e) if memory::dataset_absent(&e) => match memory {
+            Memory::Remote { uri } => Err(memory::empty_remote(uri)),
             Memory::Local { path } => Err(anyhow::anyhow!("no index found at {}", path.display())),
         },
         Err(e) => Err(e),
