+++ tests/first_index.rs
@@ -20,7 +20,7 @@ fn write_session(source: &std::path::Path) {
 }
 
 async fn chunk_count() -> usize {
-    let s = funes::commands::recall::status(funes::memory::hub::Memory::local())
+    let s = funes::commands::recall::status(funes::memory::Memory::local())
         .await
         .unwrap();
     s.lines()
