+++ tests/redact_on_index.rs
@@ -54,7 +54,7 @@ async fn planted_key_is_redacted_at_index_time() {
         .unwrap();
 
     // Read the stored turn back: the marker is present, the key body is gone.
-    let got = funes::commands::recall::get(funes::memory::hub::Memory::local(), session.into(), "t1".into(), 3)
+    let got = funes::commands::recall::get(funes::memory::Memory::local(), session.into(), "t1".into(), 3)
         .await
         .unwrap();
     assert!(
