+++ tests/reindex_incremental.rs
@@ -19,7 +19,7 @@ fn write_session(source: &std::path::Path, n_turns: usize) {
 }
 
 async fn chunk_count() -> usize {
-    let s = funes::commands::recall::status(funes::memory::hub::Memory::local())
+    let s = funes::commands::recall::status(funes::memory::Memory::local())
         .await
         .unwrap();
     s.lines()
